simple.test/my_strands_agent.py
from bedrock_agentcore import BedrockAgentCoreApp
from strands import Agent

app = BedrockAgentCoreApp()
agent = Agent()

@app.entrypoint
def invoke(payload):
    """Your AI agent function"""
    user_message = payload.get("prompt", "Hello! How can I help you today?")
    result = agent(user_message)
    return {"result": result.message}

if __name__ == "__main__":
    app.run()



# To test
# 1. python my_strands_agent.py
# 2. go to second terminal window run: 
#                                       curl -X POST http://localhost:8080/invocations   -H "Content-Type: applicatioon"   -d '{"prompt": "Hello!"}'