simple.test/requirements.txt
bedrock-agentcore
strands-agents
pytest
requests
langgraph
langchain-aws
langchain-core